% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateDataset.R
\name{annotateFootprintData}
\alias{annotateFootprintData}
\title{Annotate the Footprint Data}
\usage{
annotateFootprintData(fp.df, chipseq.hits, host = "localhost",
  port = "5432")
}
\arguments{
\item{fp.df}{The data frame of FIMO motifs and footprints}

\item{chipseq.hits}{The dataframe of ChIPseq hits corresponding to the dataset}
}
\value{
The supplied data frame with added columns for the one-hot coded TF class data, GC
content, and asinh-transformed distance from the nearest transcription start site
}
\description{
Annotate the footprint and FIMO dataframe by adding TF class, GC content, and TSS distance
}
