% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildModels.R
\name{buildBoostedModel}
\alias{buildBoostedModel}
\title{Build a Gradient Boosted Model}
\usage{
buildBoostedModel(dataList)
}
\arguments{
\item{dataList}{The list of prepared data, generally obtained by running the \code{prepModelData}
function. It must contain at least these 4  matrices: X_train, y_train, X_test, y_test}
}
\value{
The stats dataframe from the model and the model itself
}
\description{
Build the gradient boosted model for the given list of data and return the stats data frame for
inspection and plotting
}
