% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildModels.R
\name{buildLinearModels}
\alias{buildLinearModels}
\title{Build a Linear Models}
\usage{
buildLinearModels(dataList)
}
\arguments{
\item{dataList}{The list of prepared data, generally obtained by running the \code{prepModelData}
function. It must contain at least these 4  matrices: X_train, y_train, X_test, y_test}
}
\value{
A stats dataframes, containing statistics for the both thefull linear model and for
the individual linear models
}
\description{
Build a linear model with all features and linear models with each individual feature for the
given list of data and return the stats data framed for inspection and plotting
}
