% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructMotifDataset.R
\name{constructLymphoblastDataset}
\alias{constructLymphoblastDataset}
\title{Construct the Lymphoblast Datast}
\usage{
constructLymphoblastDataset()
}
\value{
The complete motif/ChIPSeq dataset for lymphoblast
}
\description{
Using the included ChIPSeq data, construct the lymphoblast dataset
}
