% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildModels.R
\name{createBiasOneMotif}
\alias{createBiasOneMotif}
\title{Create a biased training set for one motif}
\usage{
createBiasOneMotif(train.df, motif, negPosRatio = 9)
}
\arguments{
\item{train.df}{The data frame of training data}

\item{motif}{The motif of interest; this MUST be part of the dataset, else you'll get an error.}

\item{negPosRatio}{The desired ratio of negatives to positives, must be expressed as an integer
(default = 9)}
}
\value{
A reduced training set for the supplied motif, where the ratio of negatives to positives
does not exceed that supplied to the function. It may fall short if there are motifs where there
are insufficient negative points to fulfill the desired ratio, as it does not discard any
positives
}
\description{
Create a biased training set for one motif by reducing the ratio of ChIPseq non-hits to hits to a
given ratio. This should only be run on training data, as running it on testing or validation data
is not good practice.
}
