% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateDataset.R
\name{createMotifClassMap}
\alias{createMotifClassMap}
\title{Create Motif Class Map}
\usage{
createMotifClassMap(fp.df)
}
\arguments{
\item{fp.df}{The data frame of FIMO motifs and footprints. It is used for filtering out only those
motifs that appear in the data frame.}
}
\value{
The one-hot data frame of motifs and their membership in different classes
}
\description{
Create the one-hot form of the JASPAR motifs and their classes
}
