% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructMotifDataset.R
\name{createTfDf}
\alias{createTfDf}
\title{Pull all FIMO Motifs for a Given TF}
\usage{
createTfDf(TF, verbose = FALSE)
}
\arguments{
\item{TF}{A transcription factor of interest}

\item{verbose}{A Boolean flag indicating whether the function should print output (default = FALSE)}
}
\value{
The data frame containing all FIMO motifs that correspond to the supplied TF
}
\description{
Given a particular TF, pull all FIMO data for the motifs that map to that TF
}
