% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildModels.R
\name{extractMaxMCC}
\alias{extractMaxMCC}
\title{Extract the Maximum MCC Values}
\usage{
extractMaxMCC(statsDF)
}
\arguments{
\item{statsDF}{A data frame of statistics created by one of the model building functions,
\code{buildBoostedModel} or \code{buildLinearModels}. This can only be a data frame, not a
list of data frames. In order to use multiple data frames, bind them together into one data
frame first}
}
\value{
A table summarizing the maximum MCC for each model in the supplied data frame
}
\description{
Using the statistics data frames from building models, extract the maximum MCC values for
each model
}
