\name{FPML-package}
\alias{FPML-package}
\alias{FPML}
\docType{package}
\title{
\packageTitle{FPML}
}
\description{
\packageDescription{FPML}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{FPML}
\packageIndices{FPML}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{FPML}

Maintainer: \packageMaintainer{FPML}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
