% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateDataset.R
\name{getGCContent}
\alias{getGCContent}
\title{Add GC content}
\usage{
getGCContent(start_col, end_col, chrom_col, shoulder = 100)
}
\arguments{
\item{start_col}{The name of the column containing the starting location of each motif}

\item{end_col}{The name of the column containing the ending location of each motif}

\item{chrom_col}{The name of the column containing the chromosome for each motif}

\item{shoulder}{An integer indicating the distance on either side of the center of the sequence
to use as a window (default = 100)}
}
\value{
The GC content of the supplied region
}
\description{
Add GC content to a row of the data frame containing FIMO motifs and footprint data
}
