% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildModels.R
\name{joinModelData}
\alias{joinModelData}
\title{Join the Seed 16 and Seed 20 Data}
\usage{
joinModelData(full.16, full.20)
}
\arguments{
\item{full.16}{The annotated data frame containing the seed 16 data}

\item{full.20}{The annotated data frame containing the seed 20 data}
}
\value{
A joined data frame, containing footprints from both seeds. The footprint columns of this
data frame have added suffixes to denote which seed they come from, as there are now 2 of each
column, one originating from each seed size. For example, "h_frac" becomes "h_frac_20" and
"h_frac_16"
}
\description{
Join the seed 16 and seed 20 data in preparation for building models on the full dataset. This
joined data can be used to run models that either include or exclude footprint data.
}
