% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_functions.R
\name{make.pred.df.from.glm}
\alias{make.pred.df.from.glm}
\title{Create a Data Frame of Predictions for ChIPseq Data Using a Linear Model}
\usage{
make.pred.df.from.glm(glm.model, df.testdata)
}
\arguments{
\item{glm.model}{A generalized linear model that corresponds to the supplied data}

\item{df.testdata}{A data frame containing test data. It must include columns for all the
predictors, plus the actual response as the column "ChIPseq.bound"}
}
\value{
A data frame containing the actual response, the predicted response, and the name of the
supplied model
}
\description{
Given a linear model and a data frame containing labeled predictors of test data and the actual
responses as the "ChIPseq.bound" column, predict the responses and return a data frame of the
actual responses, the predicted responses, and the model name
}
