% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_functions.R
\name{make.pred.df.from.model}
\alias{make.pred.df.from.model}
\title{Create a Data Frame of Predictions for ChIPseq Data Using a Boosted Model}
\usage{
make.pred.df.from.model(gbdt.model, X.testdata, y.testdata)
}
\arguments{
\item{gbdt.model}{A gradient boosted model}

\item{X.testdata}{A matrix of predictors in the testing set}

\item{y.testdata}{A vector of responses in the testing set}
}
\value{
A data frame containing the model name, the predicted responses, and the actual responses
for the test dataset
}
\description{
Given a boosted model, a set of predictors, and a set of responses, create predicted
responses and put them together with the actual responses.
}
