% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_functions.R
\name{make.stats.df.from.preds}
\alias{make.stats.df.from.preds}
\title{Create the "Stats" Data Frame from Prediction Data Frame}
\usage{
make.stats.df.from.preds(pred.df, num.roc.points = 101)
}
\arguments{
\item{pred.df}{Data frame containing the actual reponses, predicted responses, and model name for
a given model. Generally, this should be output by either \code{make.pred.df.from.model} or
\code{make.pred.df.from.glm}}

\item{num.roc.points}{An integer indicating how many points to include in calculating the ROC
curve parameters}
}
\value{
A data frame containing statistics needed to plot the ROC and Precision-Recall curves
}
\description{
Given the "Prediction" data frame generated by a particular model type, compute the statistics
needed to plot the ROC curve, and Precision-Recall curve
}
