% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructMotifDataset.R
\name{mapTFsToMotifs}
\alias{mapTFsToMotifs}
\title{Create a TF-Motif Map in List Form}
\usage{
mapTFsToMotifs(chipseq.hits)
}
\arguments{
\item{chipseq.hits}{A table of chipseq hits}
}
\value{
A list where each name is a TF from the supplied validation set and the each entry is
a character vector of motifs that map to that TF
}
\description{
Given a set of validation data (generally ChIPseq data), return a mapping of all TFs in the data
to their JASPAR motifs. This will only return TFs for which there is a mapping to a JASPAR motif
via our TOMTOM motif matching. If there is no corresponding JASPAR motif, the TF will be dropped.
}
