% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_functions.R
\name{mattcc}
\alias{mattcc}
\title{Compute Matthews Correlation Coefficient}
\usage{
mattcc(true_classes, predicted_classes)
}
\arguments{
\item{true_classes}{A binary vector of classes that indicate the true classification of each
point}

\item{predicted_classes}{A binary vector of classes that indicate the predicted classification
of each point}
}
\value{
The Matthews Correlation Coefficient for the given prediction/truth sets
}
\description{
Given the true classes and predicted classes, compute Matthews Correlation Coefficient
}
