% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeFootprintData.R
\name{mergeFootprintsOneChrom}
\alias{mergeFootprintsOneChrom}
\title{Merge Footprints for 1 Chromosome}
\usage{
mergeFootprintsOneChrom(chrom_str, fimo_tbl, hint_regions_tbl, hint_hits_tbl,
  well_regions_tbl, well_hits_tbl)
}
\arguments{
\item{chrom_str}{A string representing one chromosome; must be one of c(1:22, "X", "Y")}

\item{fimo_tbl}{A dataframe of FIMO motifs and ChIPSeq data}

\item{hint_regions_tbl}{A database table of HINT regions}

\item{hint_hits_tbl}{A database table of HINT hits}

\item{well_regions_tbl}{A database table of Wellington regions}

\item{well_hits_tbl}{A database table of Wellington hits}
}
\value{
The original FIMO dataframe with added columns for footprint counts and the maximum or
minimum score
}
\description{
Merge HINT and Wellington footprints into the dataframe of FIMO motifs and ChipSeq for
1 chromosome
}
