% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeFootprintData.R
\name{mergeLymphoblastFootprints}
\alias{mergeLymphoblastFootprints}
\title{Merge Footprints for Lymphoblast Data}
\usage{
mergeLymphoblastFootprints(fimo.df, seedNum, host = "localhost",
  verbose = TRUE)
}
\arguments{
\item{fimo.df}{A data frame containing FIMO motifs and ChIPSeq data, generally constructed
using the \code{constructLymphoblastDataset.R} function}

\item{seedNum}{A number (either 16 or 20) denoting which seed size to use for the merge}

\item{host}{A string denoting the database host (default = "localhost")}

\item{verbose}{A Boolean value that determines whether the function should print output during
the merge (default = TRUE)}
}
\value{
The supplied FIMO/ChIPSeq data frame with added columns for HINT count, HINT maximum
score, Wellington count, and Wellington minimum score
}
\description{
Merge the HINT and Wellington footprints into the dataframe of FIMO motifs and ChIPSeq Data for
all chromosomes
}
