% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot.all.curve}
\alias{plot.all.curve}
\title{Plot All Curves}
\usage{
\method{plot}{all.curve}(stat.dataframe)
}
\arguments{
\item{stat.dataframe}{A data frame containing statistics for a particular model or group of
models. It must contain the column "threshold".}
}
\value{
A plot of all curves, where each curve corresponds to a different statistic. Note that
for this plot, the statistics are not separated by model type.
}
\description{
Plot all parameters from the statistics dataframe
}
