% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot.mattcc.curve}
\alias{plot.mattcc.curve}
\title{Plot a Matthews Correlation Coefficient Curve}
\usage{
\method{plot}{mattcc.curve}(stat.dataframe)
}
\arguments{
\item{stat.dataframe}{A data frame containing statistics for a particular model or group of
models. It must contain columns for "threshold", "MattCC", and "Model.Name"}
}
\value{
A plot of Matthews Correlations Coefficient curves for all models in the statistics
data frame
}
\description{
Plot a Matthews Correlation Coefficient Curve for the given statistics dataframe
}
