% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot.precrecall.curve}
\alias{plot.precrecall.curve}
\title{Plot Precision-Recall Curve}
\usage{
\method{plot}{precrecall.curve}(stat.dataframe)
}
\arguments{
\item{stat.dataframe}{A data frame containing statistics for a particular model or group of
models. It must contain columns for "sensitivity", "ppv", and "Model.Name"}
}
\value{
A plot of Precision-Recall curves for all models in the statistics dataframe
}
\description{
Plot a Precision-Recall curve for the given statistics dataframe
}
