% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot.roc.curve}
\alias{plot.roc.curve}
\title{Plot ROC Curve}
\usage{
\method{plot}{roc.curve}(stat.dataframe, printAUC = FALSE)
}
\arguments{
\item{stat.dataframe}{A data frame containing statistics for a particular model or group of
models. It must contain columns for "specificity", "sensitivity", and "Model.Name", else an error
will be returned}

\item{printAUC}{A Boolean variable indicating whether or not the function should print the
area under the curve on the plot (default = FALSE)}
}
\value{
A plot of ROC curves for all models in the statistics dataframe
}
\description{
Plot a ROC curve for the given statistics dataframe.
}
