% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildModels.R
\name{plotImportanceMatrix}
\alias{plotImportanceMatrix}
\title{Plot the Importance Matrix for a Gradient Boosted Model}
\usage{
plotImportanceMatrix(gbModel, X_train, filePath)
}
\arguments{
\item{gbModel}{A gradient boosted model, generally obtained using the \code{buildBoostedModel}
function}

\item{X_train}{The matrix of training data used to created the supplied model. This is generally
created as an element of the list returned by the \code{prepModelData} function}

\item{filePath}{The path to the .PNG file that will be created by the function. This file will
house the figure of the importance matrix}
}
\value{
A .PNG file containing the importance matrix figure and located at the supplied file path
}
\description{
Plot the importance matrix, showing the features with the largest gain, for a given gradient
boosted model and the full path to a desired .PNG file containing the figure
}
