% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildModels.R
\name{prepModelData}
\alias{prepModelData}
\title{Prepare Dataset for Modeling}
\usage{
prepModelData(annotated.df, seed, hintCutoff = -Inf, wellCutoff = Inf,
  motifsOnly = FALSE, biasTraining = FALSE, biasRatio = 9)
}
\arguments{
\item{annotated.df}{The fully annotated FIMO motif dataset. This must be a data frame, or else
the function will stop and give an error.}

\item{seed}{A character string that should be one of c("16","20","both"). This seed should
match the data in the data frame}

\item{hintCutoff}{A cutoff value used as a threshold on the HINT footprint score. Only scores
above this threshold will be kept (default = -Inf)}

\item{wellCutoff}{A cutoff value used as a threshold on the Wellington footprint score. Only
scores below this threshold will be kept (default = Inf)}

\item{motifsOnly}{A Boolean value indicating whether to prepare data such that it uses only motifs,
neglecting the footprint data (default = FALSE)}
}
\value{
A list in which the filtered data frame has been split into 6 pieces for X/Y and
training/testing/validation sets. These sets are completely ready to run gradient boosted
and linear models and should be passed as a list to those commands accordingly
}
\description{
Prepare the fully annotated FIMO motif dataset for running models by filtering it accordingly
}
