% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructMotifDataset.R
\name{sampleTfDataset}
\alias{sampleTfDataset}
\title{Sample the FIMO/ChIPSeq Dataset}
\usage{
sampleTfDataset(all.TF.df, sampleSize)
}
\arguments{
\item{all.TF.df}{The FIMO/ChIPSeq dataset, a data frame}

\item{sampleSize}{An integer indicating the sample size to be taken}
}
\value{
A subset of the supplied dataframe, with row number equal to sampleSize
}
\description{
Take a random sample of some size from the FIMO/ChIPSeq dataset
}
