% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateDataset.R
\name{addTSSDistance}
\alias{addTSSDistance}
\title{Add Transcription Start Site Distance}
\usage{
addTSSDistance(annotated.df, host = "localhost")
}
\arguments{
\item{annotated.df}{A dataframe containing FIMO motifs, ChIPSeq data, and footprints data}

\item{host}{A string indicating the location of the hg38 database (default = "localhost")}
}
\value{
The dataframe with an added column containing the asinh-transformed distance from each
motif to the nearest transcription start site
}
\description{
Add transcription start site distance to a dataframe of FIMO motifs and footprints data
}
