% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructMotifDataset.R
\name{constructLymphoblastDataset}
\alias{constructLymphoblastDataset}
\title{Construct the Lymphoblast Datast}
\usage{
constructLymphoblastDataset(distinctFlag = TRUE, sampleSize = NULL,
  numWorkers = 62, fimoHost = "localhost", fimoPort = "5432",
  isTest = FALSE)
}
\arguments{
\item{distinctFlag}{A Boolean flag indicating whether or not to take just the unique rows
of the data frame (default = TRUE)}

\item{sampleSize}{An integer value indicating how many rows of the data frame to take. If this
argument is not specified or make to be NULL, the entire data frame will be returned
(default = NULL)}

\item{numWorkers}{The number of parallel processes that should be run; in general, this should
either be set to the number of expected TFs or to a number restricted by the constraints of
the system (default = 62)}

\item{fimoHost}{A string indicating the location of the FIMO database (default = "localhost")}

\item{fimoPort}{A string indicating the port where the FIMO database is located (default = "5432")}

\item{isTest}{A Boolean flag indicating whether this script is being run as a test. If TRUE,
it will only use 2 TFs instead of the full set to save on time (default = FALSE)}
}
\value{
The complete motif/ChIPSeq dataset for lymphoblast
}
\description{
Using the included ChIPSeq data, construct the lymphoblast dataset
}
