% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructMotifDataset.R
\name{createTfDf}
\alias{createTfDf}
\title{Pull all FIMO Motifs for a Given TF}
\usage{
createTfDf(TF, TFs.to.motifs, chipseq.regions, chipseq.hits, verbose = FALSE)
}
\arguments{
\item{TF}{A transcription factor of interest}

\item{TFs.to.motifs}{A list where each entry is a character vector of motifs and the name of each entry
is a transcription factor}

\item{chipseq.regions}{A table of ChIPseq regions}

\item{chipseq.hits}{A table of ChIPseq hits}

\item{verbose}{A Boolean flag indicating whether the function should print output (default = FALSE)}
}
\value{
The data frame containing all FIMO motifs that correspond to the supplied TF
}
\description{
Given a particular TF, pull all FIMO data for the motifs that map to that TF
}
