% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildModels.R
\name{createThresholdPlot}
\alias{createThresholdPlot}
\title{Create a threshold plot}
\usage{
createThresholdPlot(stats.df, plotFile)
}
\arguments{
\item{stats.df}{A dataframe of statistics generated by one of the model runs}

\item{plotFile}{A path to a file for saving the generated plot as a .png file}
}
\value{
A .png file containing a plot of 3 metrics (sensitivity, specificity, ppv) as a function
of threshold
}
\description{
Create a "threshold" plot, which plots the 3 relevant prediction metrics against threshold to
enable identification of the ideal "threshold" for ChIPseq hit probability.
}
